refactor(posix/supervisor): move zcbor helpers to its own file
