+++ src/samples/supervisor/posix/ocred.c
@@ -23,6 +23,7 @@
 #include <ocre/ocre.h>
 
 #include "ipc.h"
+#include "zcbor_helpers.h"
 
 #define SOCK_PATH	   "ocre.sock"
 #define RX_BUFFER_SIZE	   2048
@@ -34,100 +35,14 @@
 /* Global ocre context for the server */
 static struct ocre_context *ctx = NULL;
 
-static void print_hex(const char *label, const uint8_t *data, size_t len)
-{
-	printf("HEX %s: ", label);
-	for (size_t i = 0; i < len; i++) {
-		printf("%02x ", data[i]);
-	}
-	printf("\n");
-}
-
-/* Helper to decode a string or nil, returns true if string was decoded, false if nil */
-static bool decode_string_or_nil(zcbor_state_t *state, char *buffer, size_t buffer_size, bool *is_nil)
-{
-	*is_nil = false;
-
-	/* Try to decode nil first */
-	if (zcbor_nil_expect(state, NULL)) {
-		*is_nil = true;
-		buffer[0] = '\0';
-		return true;
-	}
-
-	zcbor_pop_error(state);
-
-	/* Decode as string */
-	struct zcbor_string str;
-	if (!zcbor_tstr_decode(state, &str)) {
-		return false;
-	}
-
-	size_t copy_len = str.len < buffer_size - 1 ? str.len : buffer_size - 1;
-	memcpy(buffer, str.value, copy_len);
-	buffer[copy_len] = '\0';
-
-	return true;
-}
-
-/* Helper to decode a string array (CBOR list of strings) */
-static bool decode_string_array(zcbor_state_t *state, char **arr, size_t max_count, size_t *out_count, bool *is_nil)
-{
-	*is_nil = false;
-	*out_count = 0;
-
-	/* Try to decode nil first */
-	if (zcbor_nil_expect(state, NULL)) {
-		*is_nil = true;
-		return true;
-	}
-
-	zcbor_pop_error(state);
-
-	/* Decode as list */
-	if (!zcbor_list_start_decode(state)) {
-		return false;
-	}
-
-	size_t count = 0;
-	while (!zcbor_list_end_decode(state) && count < max_count) {
-		struct zcbor_string str;
-		if (!zcbor_tstr_decode(state, &str)) {
-			break;
-		}
-
-		/* Allocate and copy string */
-		arr[count] = malloc(str.len + 1);
-		if (arr[count] == NULL) {
-			return false;
-		}
-		memcpy(arr[count], str.value, str.len);
-		arr[count][str.len] = '\0';
-		count++;
-	}
-
-	arr[count] = NULL; /* NULL-terminate the array */
-	*out_count = count;
-
-	return true;
-}
-
-/* Helper to free a string array */
-static void free_string_array(char **arr, size_t count)
-{
-	for (size_t i = 0; i < count; i++) {
-		free(arr[i]);
-	}
-}
-
-/* Helper to encode a string or nil */
-static bool encode_string_or_nil(zcbor_state_t *state, const char *str)
-{
-	if (str == NULL) {
-		return zcbor_nil_put(state, NULL);
-	}
-	return zcbor_tstr_put_term(state, str, STRING_BUFFER_SIZE);
-}
+// static void print_hex(const char *label, const uint8_t *data, size_t len)
+// {
+// 	printf("HEX %s: ", label);
+// 	for (size_t i = 0; i < len; i++) {
+// 		printf("%02x ", data[i]);
+// 	}
+// 	printf("\n");
+// }
 
 static int handle_context_create_container(zcbor_state_t *dec_state, uint8_t *tx_buf, size_t tx_buf_size)
 {
@@ -881,8 +796,8 @@ int main(void)
 				break;
 			}
 
-			printf("Received %d bytes\n", n);
-			print_hex("received", rx_buf, n);
+			// printf("Received %d bytes\n", n);
+			// print_hex("received", rx_buf, n);
 
 			/* Process the request and generate response */
 			int response_len = process_request(rx_buf, n, tx_buf, sizeof(tx_buf));
@@ -892,7 +807,7 @@ int main(void)
 				break;
 			}
 
-			print_hex("sending", tx_buf, response_len);
+			// print_hex("sending", tx_buf, response_len);
 
 			if (send(s2, tx_buf, response_len, 0) < 0) {
 				perror("send");
